Fix bug related to group rights in _runas
When running a command as given system_user (eg. postgres) , _runas was only
doing a setuid but no setgid, leading to a process with only the owner rights
on files. This could result to a "permission denied" error when some file
does not belongs to the postgres user, but are supposed to be readable
by a given group. This is the case with the SSL private key under Debian which
is readable to the "ssl-cert" group.

This patch set the groups of the given system_user before setuid and
running the command. This allows the processus to have the rights given to
the groups its owner belongs to.

This should fix the second part of the issue #11 on github, reported by
ARogovsky.