Do not use crm_node --partition to discover resources
During the promote action, we used to call crm_node --partition to find other
existing nodes and compare LSNs to check if the current node to promote
is the best one. If some node in the partition does not host one of the clone,
we were failing the promotion because we couldn't check its LSN.

The call of crm_node has been replaced by some computation on
environment variables given during notify to find out which clones on
what nodes are actually taking part of the PgSQL cluster. The node list
is then passed to the promote action using a cluster private attribute
(outside of the CIB to not break the transition).

This should fix gh issue #9.