Detect and deal master/slave recovery transition
Pacemaker transition to recover a slave is composed of stop-start actions and
demote-stop-start-promote actions for a master. Before this patch, PAF
was raising a failure during the various control on the local instance
before the demote, stop or start actions. This failure usualy lead to a
fencing of the node because the RA was reported not able to stop the
local resource.

This patch tries to detect a recovery attempt from the CRM of a failing
slave or a failing master. In such cases, it gives a chance to the local
failing PgSQL instance to recover using the postgres internal recovery
process before unleashing the Pacemaker normal transition. If it
success, the actions will always detect a clean instance and go on,
avoiding a node fence AND recovering the resource.