Detect and enforce reliability of a switchover
Detect controled switchover in a transition using the notify env variables:

  * the old master node is master
  * the old master will be demoted
  * the old master will not be stopped
  * the new master is a standby
  * the new master will be promoted

In such case, after the old master is demoted and before the new master
is promoted, we check during the pre-promote action that the new master
received the "checkpoint shutdown" from the master. This allows to make
sure the new master was actually replicating with the old master when it
has been demoted.

This protect us against a race condition where the new master is
disconnected right before the old master is shutdown and did not
received the very last transactions. In such case, the transition will
be interrupted by an error during the promotion of the new master. The
new master will not be promoted and will stay in standby.