Use same role name than the system_user to connect
Before this patch, we force the role connection to "postgres" to connect
to the local instance using psql. This could lead to incompatility when
the instance was eg. created with a different system user than "postgres".

This patch just remove the connection role to use from the connstr and
leave the libpq deal with its default values (which is to use the same
name than the system user running the command). So running psql with
the same system user than the one that initdb the instance (without the
-U arg) should work as expected.

Note: fix a debug message in _query as well.

This was reported by nnn-dev in gh issue #24.