Choose a clone to promote when no master score exist
During the very first start of the cluster, no master score will
exists on any of the existing slaves, unless an admin designated
one of them using crm_master. If no master exists the cluster will
not promote a master among the slaves.
To solve this situation, we check if there is at least one master
score existing on one node in the cluster. Do nothing if at least
one master score is found among the clones of the resource. If no
master score exists, set a score of 1 only if the resource was a
shut downed master before the start.