Use given action timeout in pg_ctl when possible
We were previously using a default timeout of 10 minutes during
postgresql start or stop  using pg_ctl. This patch tries to detect the
Pacemaker's timeout for the current action and use it to set pg_ctl's
one.

The main goal here is to make sure pg_ctl does not give up before
Pacemaker. We don't want to deal with an error coming from pg_ctl where
PAcemaker is supposed to take the appropriate decisions on action
timeout.

Fix gh issue #2