Avoid fatal error during validate action when possible
Fatal error forbid the resource to run anywhere. If resource is
not installed on one node or has an error in some config file, the
previous behavior was to forbid the resource anywhere and to actually
stop everywhere.

Returning a hard failure leads to a local error which will lead to local
actions on the node for this resource, not cluster wide for all the
clones.

This was reported and the solution suggested by Benoit Lobréau during
his various tests.