Add a log message for missing binaries
The RA was raising an error without explanation when one or many
binaries where lacking.

The patch moves this test after we confirm that we have a local
instance to deal with and lower the error level so it doesn't stop all
the clone of the resource on other nodes if binaries are missing on the
current node only.