Avoid flooding log files with warning about non used nodes
PAF was issuing a warning iand set a negative score to a node each
time it does not find PostgreSQL running on that node.

Raising such a warning might be a valid concern but not raising it
during each monitor actions.

This patch avoid raising the warning message ans setting the score
more than once. Moreover, it will saves some activity between PAF and
attrd for setting a score which has been already set to -1000 from a
past monitor action.

Close #138 on github
Reported by timdeluxe