Remove managment nameserver from resolv.conf
Nameserver on eth0 hides the one on eth1 when NetworkManager set it
first. Any request for eg. srv1.priv fails as dnsmasq on eth0 does
not know IP addresses and name on eth1.

This patch just remove the nameserver from eth0 as this interface
is only needed for vagrant internal use.