Force timeout to the pg_isready command
Default timeout of command pg_isready is 3 seconds. If the serveur did not
reply within this time window, pg_isready exits with a return code of 2.

This return code doesn't mean the local PgSQL cluster was not listening, it
just means it couldn't answer within 3 seconds. However, before this patch,
PAF was considering the local cluster was not listening in such situation,
which was conflicting with pg_ctl return code, leading to an error.