Fix application_name's check
The old regex from the recovery.conf era (pre-v12 versions) has been kept
while the output format we parse is slightly different in v12. Since, v12, we
parse the output of "postgres -C primary_conninfo" were there is no quotes
around the whole value. Because of this, the check was reporting a wrong error
if application_name was the very first parameter appearing in primary_conninfo.

Thanks to @zeroone on github for testing and reporting this bug with details in
issue #156.